+++ css/designSystem.css
@@ -1,0 +1,81 @@
+[data-theme="light"] {
+    --bouton-background: #d0d0d0;
+    --bouton-background-hover: #000000;
+    --bouton-color: #000000;
+    --bouton-color-hover: #fff;
+}
+
+
+[data-theme="dark"] {
+    --bouton-background: #505050;
+    --bouton-background-hover: #fff;
+    --bouton-color: #fff;
+    --bouton-color-hover: #000000;
+}
+
+.flexBordered {
+    width: 100%;
+    height: 100%;
+    border: 2px solid var(--border-color);
+    padding: 1rem;
+    margin: 1rem;
+}
+
+.bouton {
+    background-color: var(--bouton-background);
+    color: var(--bouton-color);
+	cursor: pointer;
+	transition: background-color 0.3s, color 0.3s, border-color 0.3s;
+    border-radius: 10000000000px;
+}
+.bouton:hover {
+	background-color: var(--bouton-background-hover);
+	color: var(--bouton-color-hover);
+	border-color: white;
+}
+
+.loader {
+    margin: 0;
+    padding: 0;
+	position: absolute;
+	width: 100%;
+	height: 100%;
+}
+.loader::before, .loader::after {
+	content: "";
+	position: absolute;
+	top: 50%;
+	left: 50%;
+	border-radius: 100%;
+	animation: rotationLoader 3s infinite;
+	--translate: translate(-50%, -50%);
+	transform: var(--translate);
+}
+.loader::before {
+	height: 10rem;
+	width: 10rem;
+	background: linear-gradient(to bottom, black 0%, black 50%, white 50%, white 100%);
+	border: 4px solid black;
+}
+.loader::after {
+	height: 2rem;
+	width: 2rem;
+	background-color: white;
+	border: 4px solid black;
+}
+@keyframes rotationLoader {
+	0% {
+		transform: var(--translate) rotate(0deg);
+        left : 30%;
+	}
+    50% {
+		transform: var(--translate) rotate(720deg);
+        left : 70%;
+	}
+    /** 2 pas en avant, 1 en arrière! */
+	100% {
+		transform: var(--translate) rotate(360deg);
+        left : 30%;
+	}
+}
+
