+++ css/style.css
@@ -1,0 +1,5 @@
+@import 'designSystem.css';
+@import 'App.css';
+@import 'listeJeux.css';
+@import 'jeuSelectionne.css';
+@import 'Composants/popup.css';
