+++ index.html
@@ -3,8 +3,8 @@
 	<head>
 		<title>TP SDV CSIMAR</title>
 		<meta charset="UTF-8">
-		<link rel="stylesheet" href="style.css">
-		<script src="script.js"></script>
+		<link rel="stylesheet" href="css/style.css">
+		<script src="bundle.js"></script>
 	</head>
 	<body>
         <header class="flexBordered">
@@ -42,10 +42,10 @@
                 <p class="footerDevMail">cedric.simar@supdevinci-edu.fr</p>
             </div>
             <div class="footerGiantBombDiv">
-                <a class="footerGiantBombText" href="https://www.giantbomb.com/">Lien vers GiantBomb</a>
+                <a class="footerGiantBombText" target="_blank" href="https://www.giantbomb.com/">Lien vers GiantBomb</a>
             </div>
             <div class="footerSDVDiv">
-                <a class="footerSDVText" href="https://www.supdevinci.fr/">Lien vers Sup de Vinci</a>
+                <a class="footerSDVText" target="_blank" href="https://www.supdevinci.fr/">Lien vers Sup de Vinci</a>
             </div>
 		</footer>
 	</body>
