--- style.css
@@ -1,286 +1,0 @@
-[data-theme="light"] {
-    --background-color: #FFFFFF;
-    --text-color: #000000;
-    --border-color: #000000;
-    --input-background: #fefefe;
-    --input-color: #404040;
-    --bouton-background: #d0d0d0;
-    --bouton-background-hover: #000000;
-    --bouton-color: #000000;
-    --bouton-color-hover: #fff;
-}
-
-
-[data-theme="dark"] {
-    --background-color: #3F3F3F;
-    --text-color: #fff;
-    --border-color: #fff;
-    --input-background: #404040;
-    --input-color: #fefefe;
-    --bouton-background: #505050;
-    --bouton-background-hover: #fff;
-    --bouton-color: #fff;
-    --bouton-color-hover: #000000;
-}
-  
-
-body {
-	display: flex;
-	flex-direction: column;
-
-	margin: auto;
-	max-width: 900px;
-    background-color: var(--background-color);
-    color: var(--text-color);
-}
-
-html, body {
-	min-height: 100vh;
-	max-height: 100vh;
-}
-
-header {
-    display: inline-flex;
-    justify-content: flex-start;
-}
-
-header > div {
-    max-height: 200px;
-}
-
-.headerImgDiv {
-    display: flex;
-    height: 100%;
-}
-
-.headerImgDiv > img {
-    object-fit: scale-down;
-}
-
-main > nav {
-    display: inline-flex;
-    justify-content: space-between;
-}
-
-
-main > section {
-    display: inline-flex;
-    flex-wrap: wrap;
-    flex: 1;
-}
-
-footer {
-    display: inline-flex;
-
-    justify-content: space-between;
-    align-items: center;
-}
-
-footer > div {
-    margin: auto;
-    text-align: center;
-}
-
-footer > div > a {
-    color: var(--text-color);
-}
-
-
-.flexBordered {
-    width: 100%;
-    height: 100%;
-    border: 2px solid var(--border-color);
-    padding: 1rem;
-    margin: 1rem;
-}
-
-.sectionPage {
-    display: block;
-    text-align: center;
-    position: relative;
-    min-height: 20rem;
-}
-
-.multiGameDiv {
-    flex: 1;
-    display: flex;
-    width: 100%;
-    justify-content: space-evenly;
-}
-
-.multiGameDiv > .gameDiv {
-    width: 30%;
-    height: auto;
-}
-
-.multiGameDiv > .gameDiv > img {
-    object-fit: scale-down;
-    height: auto;
-    width: 100%;
-}
-
-.multiGameDiv > .gameDiv > div {
-    display : inline-flex;
-    justify-content: space-evenly;
-    flex-wrap: wrap;
-}
-
-.pSelectPage {
-    position: absolute;
-    top: 50%;
-    left: 50%;
-    -ms-transform: translate(-50%, -50%);
-    transform: translate(-50%, -50%);
-}
-
-.longDescRow {
-    flex-wrap: wrap;
-}
-
-input {
-    background-color: var(--input-background);
-    color: var(--input-color);
-}
-
-
-
-
-
-
-
-
-
-
-
-
-.loader {
-    margin: 0;
-    padding: 0;
-	position: absolute;
-	width: 100%;
-	height: 100%;
-}
-.loader::before, .loader::after {
-	content: "";
-	position: absolute;
-	top: 50%;
-	left: 50%;
-	border-radius: 100%;
-	animation: rotationLoader 3s infinite;
-	--translate: translate(-50%, -50%);
-	transform: var(--translate);
-}
-.loader::before {
-	height: 10rem;
-	width: 10rem;
-	background: linear-gradient(to bottom, black 0%, black 50%, white 50%, white 100%);
-	border: 4px solid black;
-}
-.loader::after {
-	height: 2rem;
-	width: 2rem;
-	background-color: white;
-	border: 4px solid black;
-}
-@keyframes rotationLoader {
-	0% {
-		transform: var(--translate) rotate(0deg);
-        left : 30%;
-	}
-    50% {
-		transform: var(--translate) rotate(720deg);
-        left : 70%;
-	}
-    /** 2 pas en avant, 1 en arrière! */
-	100% {
-		transform: var(--translate) rotate(360deg);
-        left : 30%;
-	}
-}
-
-
-.bouton {
-    background-color: var(--bouton-background);
-    color: var(--bouton-color);
-	cursor: pointer;
-	transition: background-color 0.3s, color 0.3s, border-color 0.3s;
-    border-radius: 10000000000px;
-}
-.bouton:hover {
-	background-color: var(--bouton-background-hover);
-	color: var(--bouton-color-hover);
-	border-color: white;
-}
-
-.gameDiv {
-	cursor: pointer;
-	transition: background-color 0.3s, color 0.3s, border-color 0.3s;
-}
-.gameDiv:hover {
-	background-color: black;
-	color: white;
-	border-color: white;
-}
-
-.gamePage {
-    display: block;
-}
-
-.gamePage > div {
-    display: inline-flex;
-    width: 100%;
-    justify-content: space-evenly;
-    padding: 1rem;
-}
-
-.gamePage > div > * {
-    margin: auto;
-}
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-.popup {
-    position: fixed;
-    z-index: 1;
-    left: 0;
-    top: 0;
-    width: 100%;
-    height: 100%;
-    overflow: auto;
-    background-color: rgba(0,0,0,0.4); /* Opacité */
-  }
-  
-  /* Modal Content/Box */
-  .popup-content {
-    background-color: #fefefe;
-    margin: 15% auto; /* 15% from the top and centered */
-    padding: 20px;
-    border: 1px solid #888;
-    width: 80%; /* Could be more or less, depending on screen size */
-  }
-  
-  /* The Close Button */
-  .close {
-    color: #aaa;
-    float: right;
-    font-size: 28px;
-    font-weight: bold;
-  }
-  
-  .close:hover,
-  .close:focus {
-    color: black;
-    text-decoration: none;
-    cursor: pointer;
-  }
